api/requirements.txt
Flask==3.0.3
Flask-Mail==0.9.1
Jinja2==3.1.3
python-dotenv==1.0.1 